% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckmr2colony.R
\name{afreq_str}
\alias{afreq_str}
\title{prints out the integer allele names and their freqs into alternating
strings in a vector, so all you have to do is cat it with sep = "backslash n" to
get it to print to the colony file.}
\usage{
afreq_str(CKF)
}
\description{
prints out the integer allele names and their freqs into alternating
strings in a vector, so all you have to do is cat it with sep = "backslash n" to
get it to print to the colony file.
}
