% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckmr-class.R
\name{ckmr_class}
\alias{ckmr_class}
\title{constructor function for the ckmr class}
\usage{
ckmr_class(L)
}
\arguments{
\item{L}{the list that you want to turn into a \code{ckmr} object.}
}
\description{
This is a constructor function and where we will document the \code{ckmr} class.
The \code{ckmr} class is an object that includes all the stuff necessary to
sample genotypes of pairs of individuals at different loci.  Primarily this
will be the object returned by \code{\link{ckmr_create}}. A \code{ckmr} object
is a named list.  At the first level are components \code{orig_data} (the original
data frame of markers that went into its construction) and \code{loci}, a component in which
each component corresponds to a locus and has the following
components
\describe{
  \item{freqs}{A vector of allele frequencies, named by the alleles.}
  \item{X_l}{A list, named by the pairwise relationships investigated, of matrices
  giving the probabilities of pairwise genotypes between two individauls having that
  relationship. }
  \item{C_l}{A matrix of probabilities giving the conditional probability of observing
  each possible genotype, given the true underyling state of the genotype in an individual.
  This is what encodes the genotyping error model that is \emph{assumed} for computing the Q ratios
  (See the paper).}
  \item{C_l_true}{A matrix of probabilities giving the conditional probability of observing
  each possible genotype, given the true underyling state of the genotype in an individual.
  This is what encodes the genotyping error model that is \emph{actually the true model} and which
  is used for simulating the genotypes when computing the Q ratios
  (See the paper).}
  \item{Y_l}{The assumed probability of observing pairwise genotypes between two individuals.  This is a
  list of matrices, one for each relationship, as named.  The quantities are obtained by
  matrix multiplication of \code{X_l} to \code{C_l}.}
  \item{Y_l_true}{The true probability of observing pairwise genotypes between two individuals.  This is a
  list of matrices, one for each relationship, as named.  The quantities are obtained by
  matrix multiplication of \code{X_l} to \code{C_l_true}.}
}
}

