% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckmr-class.R
\name{ckmr_relats}
\alias{ckmr_relats}
\title{return a vector of the relationships in a ckmr_class object}
\usage{
ckmr_relats(CK)
}
\arguments{
\item{CK}{a ckmr object}
}
\description{
Just a simple convenience function
}

