% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_integer_genotype_matrix.R
\name{create_integer_genotype_matrix}
\alias{create_integer_genotype_matrix}
\title{turns long format genotypes into a matrix of integer-code genotypes}
\usage{
create_integer_genotype_matrix(LG, AF)
}
\arguments{
\item{LG}{the genotypes in long format.  It must have the columns
Indiv (unique IDs of the individuals), Locus, gene_copy (must be 1 or 2
denoting which of the two gene copies in a diploid each allele is), and
Allele, which must be a character. If there are any missing genotypes
in the data frame, they must appear as NAs in the Allele column.}

\item{AF}{The data frame of allele frequencies that was passed
to create_ckmr() to create the ckmr object that will be used for computing
genotype probabilities.  It must have columns
Chrom, Pos, Locus, Allele, LocIdx, AlleIdx, and Freq.}
}
\description{
turns long format genotypes into a matrix of integer-code genotypes
}
