% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_L_microsat}
\alias{example_L_microsat}
\title{An example microsatellite locus}
\format{
An object of class \code{list} of length 2.
}
\usage{
data(example_L_microsat)
}
\description{
These data are from Locus SFOC113 from Ruzzante et al.  It is included
mostly to show how one might make a microsatellite genotyping error model.
This is a list with two components, each of them vectors:
\describe{
\item{freqs}{The frequencies of the different alleles. The names are the
allele lengths.}
\item{geno_freqs}{The expected frequencies of the possible genotypes given
the allele freqs. The names are the names of the genotypes, which are the
allele lengths separate by " / ".}
}
}
\keyword{datasets}
