% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_logls.R
\name{extract_logls}
\alias{extract_logls}
\title{Extract log-likelihood ratios and associated values from a Qij object}
\usage{
extract_logls(Q, numer, denom)
}
\arguments{
\item{Q}{a Qij object.  Like that returned by \code{\link{simulate_Qij}}}

\item{numer}{a named vector of weights to be given to the relationships for the numerator.
The names are the relationships and the values are the weights.  Relationships that
appear in the "Tos" field of the Q that do not appear will get weights of 0.}

\item{denom}{a named vector of weights to be given to the relationships for the denominator.
The names are the relationships and the values are the weights.  Relationships that
appear in the "Tos" field of the Q that do not appear will get weights of 0.}
}
\description{
This function applies the prior weights on the pairs for the numerator (for generality)
and denominator and returns the log-likelihood ratios in a tidy format
}
