% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mendel-interface.R
\name{fast_mendel_outped2genos}
\alias{fast_mendel_outped2genos}
\title{read genotypes of target individuals out of Mendel output pedigree file.}
\usage{
fast_mendel_outped2genos(OutPed, Def)
}
\arguments{
\item{OutPed}{path to the mendel output pedigree file}

\item{Def}{path to the mendel defs file.  This is used to get the number of alleles at
each locus, so, the model loci in the mendel run must be equal to the loci listed in Def.}
}
\value{
Returns a tbl_df with four columns: 1) Rep, a base-1 index of which simulated replicate
it is; 2) Indiv, the index of the target individual (typically 1 or 2); 3) Locus, the index
of the locus; 4) Geno, the genotype of the individual as a single number.
}
\description{
this is a version that uses awk and sed that I am sure will be much faster than
the standard R readLines approach.  But only works if those tools are installed
on the system.
}
\examples{
# first prepare files for mendel simulation and run it in tmpDir
# with naming prefix "mendel-example"
example(run_mendel)

# then grab the results
results <- fast_mendel_outped2genos(file.path(tmpDir, "mendel-example-Ped.out"), file.path(tmpDir, "mendel-example-Def.in"))

# finally show the results:
results
}

