% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flatten-ckmr.R
\name{flatten_ckmr}
\alias{flatten_ckmr}
\title{flatten a from/to pair from a ckmr object (see \code{\link{ckmr_class}}into a long vector of probabilities}
\usage{
flatten_ckmr(CK, rel, useYl_true = FALSE, unname = TRUE)
}
\arguments{
\item{CK}{an object of class ckmr}

\item{rel}{The string name of the relationship you want to extract probs for.}

\item{useYl_true}{set to TRUE if you want to use the Y_l_true field instead of Y_l.}

\item{unname}{if TRUE (the default) this strips the names off the results.}
}
\description{
blah
}

