% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckmr-class.R
\name{format.ckmr}
\alias{format.ckmr}
\title{format method for ckmr class (to print)}
\usage{
\method{format}{ckmr}(x, ...)
}
\arguments{
\item{x}{an object of class \code{\link{ckmr_class}}.}

\item{...}{additional arguments to format. (But nothing implemented).}
}
\description{
Just prints relevant information for a quick look.
}
