% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qij-class.R
\name{format.Qij}
\alias{format.Qij}
\title{format method for Qij class (to print)}
\usage{
\method{format}{Qij}(x, ...)
}
\arguments{
\item{x}{an object of class \code{\link{Qij_class}}.}

\item{...}{additional arguments to format. (But nothing implemented).}
}
\description{
Just prints relevant information for a quick look.
}
