% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ge_model_microhap1.R
\name{ge_model_microhap1}
\alias{ge_model_microhap1}
\title{create matrix C for probability of observed genotypes from microhaplotype data}
\usage{
ge_model_microhap1(
  L,
  miscall_rate = 0.005,
  dropout_rate = 0.005,
  per_snp_rate = FALSE
)
}
\arguments{
\item{L}{an element of the list created by \code{\link{long_markers_to_X_l_list}}. Such
an element basically holds the information at a single locus.  The idea here
is that every ge_mod_* function takes in an object like L, and then can
use any piece of information in it about alleles or genotypes to
configure a genotyping error model.}

\item{miscall_rate}{The rate at which microhaplotype alleles are mis-called.  If
option \code{per_snp_rate} is TRUE, then this is the rate at which each SNPs might
get miscalled, such that the pverall miscall rate for microhaplotypes with more SNPs
is higher than for microhaplotypes with few SNPs.}

\item{dropout_rate}{Rate of allelic dropout.}

\item{per_snp_rate}{Logical.  If true, then the overall mis-call rate for the
microhaplotype locus is \code{miscall_rate} times the number of SNPs in the microhapotype.
The default is FALSE, and it this option is not really recommended.}
}
\description{
This is intended for the case where the genotypes in question are composed of alleles
that are actually the multi-SNP haplotypes obtained from next generation sequence data.
In other words, all the SNPs occur on a single read and the phase is known because they
are all together on the read.  It allows for a per-locus or a per-SNP-specific sequencing error rate.  The
haplotypes must be named as strings of A, C, G, or T, (though they could be strings of
any characters---the function isn't going to check that!) and for now  we assume that if the
SNPs are multiallelic then genotyping errors to either of the alternate alleles are
equally likely.  Currently assumes that genotyping errors are equally likely in either
direction at a SNP, too.
}
\examples{
# here is some example information about a microhap:
example_L_microhap

# now we can feed it in to the function with default parameter values.
}
