% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ge_model_microsat1.R
\name{ge_model_microsat1}
\alias{ge_model_microsat1}
\title{a simple "length-aware" genotyping error model for microsatellites}
\usage{
ge_model_microsat1(L, miscall_rate = 0.005, miscall_decay = 0.8,
  dropout_rate = 0.01, dropout_scale_factor = 0.25)
}
\arguments{
\item{L}{an element of the list created by \code{\link{long_markers_to_X_l_list}}. Such
an element basically holds the information at a single locus.  The idea here
is that every ge_mod_* function takes in an object like L, and then can
use any piece of information in it about alleles or genotypes to
configure a genotyping error model. In the case of microsatellites,
the names of the alleles must be the allele lengths, like "64" or "112".}

\item{miscall_rate}{the rate at which alleles are miscalled.}

\item{miscall_decay}{the parameter (or a geometric distribution) that determines
how quickly the probability of the miscall being to a certain allele length
decreases as you get further and further away from the true allele length.
The larger this values, the more quickly the probability decays.  For every
microsatellite allele step size further from the true allele size, the probability
increases by a factor of 1 - miscall_decay.  Thus, if miscall_decay = 1
then all errors are a step-size of one away, and if it is 0, then
the miscall is equally likely to occur to any other allele length.}

\item{dropout_rate}{the base rate at which alleles of "typical" size dropout.}

\item{dropout_scale_factor}{the rate of drop out increases for alleles
larger than the (frequency-weighted) average size of allele by a factor of
dropout_scale_factor times the z-score of the allele length.}
}
\description{
In this genotyping error model, mis-calls are more likely to nearby
allele lengths than distant ones, and larger alleles are more likely
to drop out than small or average-sized ones.
}
