% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geno_err_models.R
\name{hap_obs_prob}
\alias{hap_obs_prob}
\title{a simple helper function}
\usage{
hap_obs_prob(H1, H2, e)
}
\arguments{
\item{H1}{the DNA bases at the SNPs in the first haplotype}

\item{H2}{the DNA bases at the SNPs in the second haplotype}

\item{e}{per snp error rates}
}
\description{
This will ultimately not be exported. It gives the probability that
H2 is observed from a haplotype that is truly H1 given a super
simple model of independent sequencing errors. This is not
intended to be vectorized, at the moment.
}

