% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mendel-interface.R
\name{id_prepend}
\alias{id_prepend}
\title{prepend an ID to a mendel input or output file}
\usage{
id_prepend(ID, suff)
}
\arguments{
\item{ID}{The ID to be prepended.}

\item{suff}{The final part of the file name (like Def.in, etc.)}
}
\description{
This is just a little helper fucntion to ensure consistency in file naming.
}

