% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genotypes_and_error.R
\name{index_ab}
\alias{index_ab}
\title{a function to compute the genotype index from alleles a and b}
\usage{
index_ab(a, b, A)
}
\arguments{
\item{a}{The first allele in the genotype}

\item{b}{The second allele in the genotype}

\item{A}{the total number of alleles at the locus in the population.}
}
\description{
This implements the function to index genotypes so that we can put
them into matrices.  The paper mentions this indexing, in the section
"Implementation."
}
\details{
If a is not an allele with index lower than b, then this function
automatically reorients the alleles so that it is. This is vectorized and
so should work over long vectors of alleles.
}
\examples{
# Create some data to test/demonstrate this:
A <- 7
all_genos <- expand.grid(1:A, 1:A)[,c(2,1)] \%>\%
 setNames(c("a", "b"))

# when a < b
right_ord <- all_genos \%>\%
 dplyr::filter(a <= b) \%>\%
dplyr::mutate(Idx = 1:length(a),
      Indab = index_ab(a, b, A))

# when a > b
wrong_ord <- all_genos \%>\%
 dplyr::filter(a >= b) \%>\%
 arrange(b, a) \%>\%
dplyr::mutate(Idx = 1:length(a),
      Indab = index_ab(b, a, A))

# then check to make sure it all checks out
all(right_ord$Idx == right_ord$Indab)
all(wrong_ord$Idx == wrong_ord$Indab)
}
