% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/genotypes_and_error.R
\name{insert_Y_l_matrices}
\alias{insert_Y_l_matrices}
\title{compute the Y_l matrices for each locus in a list}
\usage{
insert_Y_l_matrices(L)
}
\arguments{
\item{L}{the list of X_l and C_l matrices}
}
\description{
Once you have the X_l matrices for each kappa, and the C_l matrices
have been inserted into the list, then this function cycles over the loci
and the kappa values and does the matrix multiplication to give you the
matrix Y_l.
}
\examples{
example(insert_C_l_matrices)
mh_yl_example <- insert_Y_l_matrices(mh_cl_example)
}

