% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{linked_gene_drop_true_genotypes}
\alias{linked_gene_drop_true_genotypes}
\title{Gene-drop true linked genotypes for the two focal individuals in a pedigree}
\usage{
linked_gene_drop_true_genotypes(
  kid,
  pa,
  ma,
  n_alleles,
  map_pos,
  chrom_id,
  chrom_start,
  chrom_end,
  freqs,
  reps,
  min_crossovers = 0L
)
}
\arguments{
\item{kid}{integer vector of pedigree IDs.}

\item{pa}{integer vector of paternal IDs, with 0 for founders.}

\item{ma}{integer vector of maternal IDs, with 0 for founders.}

\item{n_alleles}{integer vector with the number of alleles at each locus.}

\item{map_pos}{numeric vector of map positions in centiMorgans.}

\item{chrom_id}{integer chromosome ID per locus.}

\item{chrom_start}{one-based start locus index for each chromosome.}

\item{chrom_end}{one-based end locus index for each chromosome.}

\item{freqs}{list of allele-frequency vectors, one per locus.}

\item{reps}{number of replicate gene drops.}

\item{min_crossovers}{minimum number of crossovers in the typed marker span
of each chromosome for each meiosis.}
}
\value{
A list with integer matrices \code{indiv1} and \code{indiv2}.
}
\description{
Gene-drop true linked genotypes for the two focal individuals in a pedigree
}
