% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linked-marker-cpp-inputs.R
\name{linked_marker_cpp_inputs}
\alias{linked_marker_cpp_inputs}
\title{prepare linked marker inputs for the Rcpp gene dropper}
\usage{
linked_marker_cpp_inputs(df, cM_per_Mb = 1)
}
\arguments{
\item{df}{A data frame in the format of \code{\link{long_markers}}.}

\item{cM_per_Mb}{Recombination-rate conversion used to create a Map column
from Pos when Map is not present. The value is centiMorgans per megabase.}
}
\description{
prepare linked marker inputs for the Rcpp gene dropper
}
\keyword{internal}
