% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{linked_mhaps}
\alias{linked_mhaps}
\title{Microhaplotype data from kelp rockfish \emph{Sebastes atrovirens} with pretend linkage positions}
\format{
This tibble includes 825 alleles/haplotypes from 165 genomic regions and
is in the format of \code{\link{long_markers}}. However since true positions
of these markers in the genome are not known I just pretended that we knew some (randomly generated)
locations.   The column
\code{Chrom} Is randomly given  \code{Locus} has the name of each locus in it.  The column
\code{Pos} has a random position (in base pairs) for each locus.  The
\code{Allele} column gives the sequences of each allele/haplotype observed in the population.
}
\usage{
data(linked_mhaps)
}
\description{
These are data from 165 amplicons sequenced in two different runs of an Illumina
MiSeq machine at the SWFSC lab in Santa Cruz.  Genotypes of roughly 150 individuals
were called an resolved into haplotypes using the fact that each short read comes
from a single chromosome. These were randomly assigned to some positions on some random
chromosomes for testing linkage stuff.
}
\keyword{datasets}
