% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{locus_specific_pairwise}
\alias{locus_specific_pairwise}
\title{Return locus-specific pairwise relationship measures between desired pairs of individuals}
\usage{
locus_specific_pairwise(S, T, s, t, values, nGenos, Starts)
}
\arguments{
\item{S}{"source", a matrix whose rows are integers, with NumInd-source rows and NumLoci columns, with each
entry being a a base-0 representation of the genotype of the c-th locus at the r-th individual.
These are the individuals you can think of as parents if there is directionality to the
comparisons.}

\item{T}{"target",  a matrix whose rows are integers, with NumInd-target rows and NumLoci columns, with each
entry being a a base-0 representation of the genotype of the c-th locus at the r-th individual.
These are the individuals you can think of as offspring if there is directionality to the
comparisons.}

\item{s}{a vector of base-1 indexes of the source individual in each pair.}

\item{t}{a vector of base-1 indexes of the target individual in each pair.  This vector is parallel to s.  So,
for example (s\link{i}, t\link{i}) designates a pair that you wish to investigate (individual s\link{i} in S and t\link{i} in T)}

\item{values}{the vector of genotype specific values.  See the probs field of \code{\link{flatten_ckmr}}.}

\item{nGenos}{a vector of the number of genotypes at each locus}

\item{Starts}{the base0 indexes of the starting positions of each locus in probs.}
}
\value{
a data frame with columns "indS" (the base-1 index of the individual in S),
"indT" (the base-1 index of the individual in S), "locus" (base-1 index of the locus),
and "value" (the value extracted, typically a log likelihood ratio).  If the pair is missing that
locus it is given as NA_REAL
}
\description{
The idea here is that you can go back and look more closely at the log-likelihood ratios
for pairs that are found to look the PO, etc., to see how much each of the different
loci are contributing.  More explanation later.
}
