% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_missing_data_array.R
\name{make_miss_mask_mat}
\alias{make_miss_mask_mat}
\title{make a miss_mask_mat for simulate_Qij()}
\usage{
make_miss_mask_mat(MG, L, reps)
}
\arguments{
\item{MG}{a matrix of N rows and L columns with a 0 if the data are missing
and a 1 otherwise.}

\item{L}{a list with two elements, r and c, each one a vector of the same length.
Each pair of r,c values gives a pair of individuals (rows in MG) whose missing
data patterns are to be used.}

\item{reps}{Minimum number of Monte Carlo reps to do.}
}
\value{
A miss_mask_mat: A logical matrix with L columns and R rows. The (r,c)-th
element is TRUE if the c-th locus should be considered missing in the r-th
simulated sample. Note that the number of rows is the number of Monte Carlo
reps that should be done.
}
\description{
This uses the missing data patterns at pairs of individuals
to specify the missing data patterns.
}
