% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mendel-interface.R
\name{markers2mendel_def_lines}
\alias{markers2mendel_def_lines}
\title{convert a long-format data frame of loci into a vector of lines for a Mendel def file}
\usage{
markers2mendel_def_lines(df)
}
\arguments{
\item{df}{A data frame in the format of \code{\link{long_markers}}.}
}
\description{
This makes a vector of text that has carriage returns and things in it as necessary.
So, to include it in a def file, if the return from this function is called
D you can just do cat(D, sep = "", file = myDef.txt), for example.
}
\details{
Because it appears that Mendel requires unique locus names we will name them
Chrom-underscore-Loc.
}
\examples{
data(long_markers)
D <- markers2mendel_def_lines(long_markers)
cat(D[1:100], sep = "")
}

