% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mendel-interface.R
\name{markers2mendel_map_lines}
\alias{markers2mendel_map_lines}
\title{create a vector of Mendel formatted map file strings}
\usage{
markers2mendel_map_lines(df)
}
\arguments{
\item{df}{A data frame in the format of \code{\link{long_markers}}.}
}
\description{
This makes a vector of text that has carriage returns and things in it as necessary.
So, to include it in a map file, if the return from this function is called
D you can just do cat(D, sep = "", file = myMap.txt), for example.
}
\examples{
data(long_markers)
D <- markers2mendel_map_lines(long_markers)
cat(D[1:100], sep = "")
}

