% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{markers_on_map}
\alias{markers_on_map}
\title{An example of how to specify markers mapped onto chromosomes}
\format{
A named list of chromosomes upon which mapped markers are available.  The component
corresponding to a chromome is itself a list which may or may not be named.  If it is named
then its names are the names of the loci. The loci should be in sorted order along the chromosome.
Each locus within a chromosome is represented as a list with the following components.
\describe{
\item{pos}{Position in centiMorgans along the chromosome. This is a little nonstandard, but you
can think of the first locus as having a certain rate of recombination with the very beginning of
the chromosome, and then pos is the cumulative sum of the inter-locus recombination fractions. }
\item{freqs}{The allele frequencies at the locus}
}
}
\usage{
data(markers_on_map)
}
\description{
This is a named list.  Each component is named with the number of the chromosome.
In this case, those numbers are: "4", "9", "14", "19", and "21". Each  component
is itself a named list that holds the position in centiMorgans and the allele frequencies
of the alleles at the marker.
}
\keyword{datasets}
