% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mendel-interface.R
\name{mendel_outped2genos}
\alias{mendel_outped2genos}
\title{slurp up the pedigree output file from a Mendel genedrop and get the needed genotypes}
\usage{
mendel_outped2genos(Pedfile)
}
\arguments{
\item{Pedfile}{The name of the pedigree output file}
}
\description{
This assumes that the target individuals are numbered 1 and 2, and it pitches the data
for everyone else, then it converts the genotypes into a big long format data frame.
}

