% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mendel-interface.R
\name{mendelBin}
\alias{mendelBin}
\title{Given the operating system, check for mendel program in the typical location}
\usage{
mendelBin()
}
\description{
Given the operating system, check for mendel program in the typical location
}

