% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_Qij.R
\name{merge_Qij}
\alias{merge_Qij}
\title{Merge the results from multiple Qij objects}
\usage{
merge_Qij(Qlist)
}
\arguments{
\item{Qlist}{A list of Qij objects}
}
\value{
A Qij object with component names identical to that of the first Qij object
on the list, with values that are the concatenated versions of all of the
Qij values in the list. The attributes of the return item are the same as the
first Qij object on the list.
}
\description{
This is a simple function that you will likely only ever need to use when
you have split up the simulation burden to do many reps of the same scenarios,
typically when you are parallelizing over different pairs of individuals
with the same amount of missing data.  It is up to the user to make sure that
this is used intelligently. i.e., don't merge Qij objects that were simulated
in different fashions (i.e. linked and unlinked).
}
