% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_kin_logl_ratios.R
\name{pairwise_kin_logl_ratios}
\alias{pairwise_kin_logl_ratios}
\title{compute pairwise log likelihood ratios between two sets of individuals}
\usage{
pairwise_kin_logl_ratios(
  D1,
  D2,
  CK,
  numer,
  denom = NULL,
  keep_top = NULL,
  num_cores = parallel::detectCores()
)
}
\arguments{
\item{D1}{long format genotype data frame of the first set of individuals.  If doing
parent-offspring, this should be the parents.}

\item{D2}{long format genotype data frame of the second set of individuals. If doing
parent-offspring, this should be the offspring.}

\item{CK}{the ckmr object created previously from the data that D1 and D2 came from.}

\item{numer}{the relationship that should be the numerator of the logl ratio.
This relationship must appear in the ckmr object CK.}

\item{denom}{the relationship that should be the denominator of the logl ratio.
This relationship must appear in the ckmr object CK. Leave empty to only calculate the
log likelihood of of the numerator.}

\item{keep_top}{If this is given as an integer, then for each individual in D2, only the keep_top
individuals with the highest logl ratios are retained for the result.  The rest
are discarded.}

\item{num_cores}{The number of cores to use with mclapply.  This is ignored
if using windows (in which case it is always set to one.)  Otherwise, by
default it uses all the cores available.}
}
\description{
This computes log likelihoods for all pairs formed by comparing every individual
in D1 to every individual in D2.  If D1 and D2 are different objects, it is assumed
that they include disjoint groups of individuals, and all pairs are returned. If
D1 and D2 are the same object, then the output is filtered so that each
pairwise comparison appears only once, and so that the comparisons between the
same individual are dropped.
}
