% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/morgan-interface.R
\name{ped2morgan_input}
\alias{ped2morgan_input}
\title{convert a pedigree into control files used by morgan to drop genes}
\usage{
ped2morgan_input(ped, markers, seeds, Dir = ".")
}
\arguments{
\item{ped}{A data frame with five columns: Kid, Pa, Ma, Sex, Observed.
The entries in the first three columns are strings or numbers, and in the
remaining two are integers: Sex: 1 - male, 2 - female, 0 - unknown.
Observed: 1 - observed, 2 - not observed. Note the founders in the pedigree
have to appear as kids in a record that has 0 and 0 for the Pa and Ma.}

\item{markers}{A list of chromosomes with markers upon them in the format of
\code{\link{markers_on_map}}}

\item{Dir}{the directory where the output files should be placed.}

\item{seed}{a vector of 2 integers, each of them positive but less than 5e08  The
first one has to be odd.}
}
\description{
convert a pedigree into control files used by morgan to drop genes
}
\examples{
   data(pedigrees)
   data(markers_on_map)
   ped2morgan_input(pedigrees$FS, markers_on_map, c(12345, 789623), "/tmp")
}

