% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mendel-interface.R
\name{pedigree2mendel_ped_file}
\alias{pedigree2mendel_ped_file}
\title{for making a pedigree file with no observed genotypes on it}
\usage{
pedigree2mendel_ped_file(df, name, filename = NA)
}
\description{
for making a pedigree file with no observed genotypes on it
}
\keyword{internal}
