% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckmr-class.R
\name{print.ckmr}
\alias{print.ckmr}
\title{print method for ckmr class}
\usage{
\method{print}{ckmr}(C)
}
\arguments{
\item{C}{an object of class \code{\link{ckmr_class}}.}
}
\description{
Just wraps a call to the format.ckmr function
}
