% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qij-class.R
\name{print.Qij}
\alias{print.Qij}
\title{print a Qij object nicely}
\usage{
\method{print}{Qij}(x, ...)
}
\arguments{
\item{x}{an object of class \code{\link{Qij_class}}}

\item{...}{additional arguments to print. (But nothing implemented).}
}
\description{
For nice printing.
}
