% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qij-class.R
\name{print_calc_relats}
\alias{print_calc_relats}
\title{A simple function to print out the calc_relats in a list-oriented fashion}
\usage{
print_calc_relats(Q)
}
\arguments{
\item{Q}{an object of class \code{\link{Qij_class}}}
}
\description{
A simple function to print out the calc_relats in a list-oriented fashion
}
