% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qij-class.R
\name{Qij_class}
\alias{Qij_class}
\title{Constructor function for the class Qij}
\usage{
Qij_class(Q)
}
\arguments{
\item{Q}{the list of lists to be turned into a Qij object.}
}
\description{
The Qij class holds the output of simulated genotype pairs whose
probabilities have been evaluated under different pairwise relationship
hypotheses.  It is output by the function \code{\link{simulate_Qij}}.
}
\details{
A Qij object is quite simply a list of lists of vectors.  At the first
level the components are named by relationships from which data were
simualated (the "froms").  Within each of those is a named list of
relationships for which the probability of each simulated genotype
pair was evaluated (the "tos").
}

