% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{read_mendel_outped}
\alias{read_mendel_outped}
\title{pick the genotypes out of the Mendel output pedigree file to use to compute Q}
\usage{
read_mendel_outped(Input, NumA, verbose)
}
\arguments{
\item{Input}{the path to the Mendel output file to read in.}

\item{NumA}{the number of alleles at each locus}

\item{verbose}{integer flag.  1 gives verbose output. 0 does not.}
}
\description{
Not sure exactly how I am going to do this, as we need to include genotyping error on there as well.
Crap.  But at least I have a start on how to parse this monstrous file. This assumes that the focal pair
of individuals are labeled 1 and 2.
}
\examples{
\dontrun{
read_mendel_outped("/Users/eriq/Desktop/mendel-example-Ped.out")
}
}
