% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/markers.R
\name{reindex_markers}
\alias{reindex_markers}
\title{re-index loci and or alleles  given a data frame of markers}
\usage{
reindex_markers(M)
}
\arguments{
\item{M}{a data frame of markers.}
}
\description{
The data frame should be of the format of \code{\link{long_markers}}.
This function will reset the values of AlleIdx and LocIdx in case
some markers (or some alleles) were removed.  It also rescales the
allele frequencies to sum to one, and, of course, sorts the loci into
map order along the chromosomes.
}
\examples{
# subsample alleles from the long_markers data frame.
# Note, you would never subsample alleles like this but it
# helps to demonstrate the function.
set.seed(5);  # for reproducibility
A_few_markers <- long_markers \%>\%
  dplyr::sample_n(1000) \%>\%     # subsample 4000 alleles
  dplyr::group_by(Chrom, Locus) \%>\%
  dplyr::mutate(NumAlle = n()) \%>\%
  dplyr::filter(NumAlle > 1)  \%>\%  # chuck out loci only one allele now...
  dplyr::select(-NumAlle)  # get rid of the NumAlle column we created

  reindex_markers(A_few_markers)   # then reindex them
}
