% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{samp_from_mat}
\alias{samp_from_mat}
\title{Sample 1 observation from cell probabilities that are columns of a matrix}
\usage{
samp_from_mat(M)
}
\arguments{
\item{M}{a matrix whose rows are reals, each summing to one}
}
\value{
a vector length = \code{nrow(M)} of indices, with each element being
the column that was chosen in that row's sampling
}
\description{
Takes a matrix in which rows sum to one. For each row, performs a
single multinomial draw from amongst the columns, weighted by their values in that column
}
