% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-linked-genotype-pairs.R
\name{sample_linked_genotype_pairs}
\alias{sample_linked_genotype_pairs}
\title{simulate genotype-pairs from linked markers}
\usage{
sample_linked_genotype_pairs(
  df,
  ped,
  C = NULL,
  num = 1000,
  useMendel = FALSE,
  cM_per_Mb = 1,
  min_crossovers = 0
)
}
\arguments{
\item{df}{A data frame in the format of \code{\link{long_markers}}.}

\item{ped}{The pedigree to be simulating from}

\item{C}{a list whose elements contain, at a minimum, the "C-matrices" which give the
probability of observed genotypes given true genotypes.  If this is NULL (the default), then the
function will assume no genotyping error.}

\item{num}{Number of reps of gene-dropping to do. Default is 1000}

\item{useMendel}{If TRUE, use the external Mendel program for linked gene
dropping. If FALSE, use CKMRsim's internal Rcpp simulator.}

\item{cM_per_Mb}{Recombination-rate conversion used to create a Map column
from Pos when Map is not present. The value is centiMorgans per megabase.}

\item{min_crossovers}{Minimum number of crossovers in the typed marker span
of each chromosome for each meiosis.}
}
\value{
This returns a list named by Chrom.Locus.Pos (the names of C), in which each
component is a vector of length num that is the integer index of the simulated pair of genotypes.
}
\description{
This is the main function to use.  Pass it a data frame of markers (indexed in order)
and it will drop the genes, do genotyping error and then return
a list of vectors which hold the genotype index that you can use to subscript the
joint probability vectors with.  This function assumes that the loci in df are ordered
appropriately (i.e have been run through reindex_markers()) and that the components in
C are named Chrom.Locus.Pos, as is typical.  Obviously the C list should correspond
exactly to the markers/alleles in df.
}
