% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-linked-true-genotype-pairs.R
\name{sample_linked_true_genotype_pairs}
\alias{sample_linked_true_genotype_pairs}
\title{simulate true genotype-pairs from linked markers using Rcpp}
\usage{
sample_linked_true_genotype_pairs(
  df,
  ped,
  num = 1000,
  cM_per_Mb = 1,
  min_crossovers = 0
)
}
\arguments{
\item{df}{A data frame in the format of \code{\link{long_markers}}.}

\item{ped}{The pedigree to be simulating from.}

\item{num}{Number of reps of gene-dropping to do. Default is 1000.}

\item{cM_per_Mb}{Recombination-rate conversion used to create a Map column
from Pos when Map is not present. The value is centiMorgans per megabase.}

\item{min_crossovers}{Minimum number of crossovers in the typed marker span
of each chromosome for each meiosis.}
}
\description{
simulate true genotype-pairs from linked markers using Rcpp
}
\keyword{internal}
