% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mendel-interface.R
\name{sample_linked_true_genotype_pairs_mendel}
\alias{sample_linked_true_genotype_pairs_mendel}
\title{simulate true genotype-pairs from linked markers using Mendel}
\usage{
sample_linked_true_genotype_pairs_mendel(df, ped, num = 1000)
}
\arguments{
\item{df}{A data frame in the format of \code{\link{long_markers}}.}

\item{ped}{The pedigree to be simulating from.}

\item{num}{Number of reps of gene-dropping to do. Default is 1000.}
}
\description{
simulate true genotype-pairs from linked markers using Mendel
}
\keyword{internal}
