% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sprinkle_markers_into_genomes.R
\name{sprinkle_markers_into_genome}
\alias{sprinkle_markers_into_genome}
\title{Randomly distribute markers along the chromosomes of a genome}
\usage{
sprinkle_markers_into_genome(markers, genome)
}
\arguments{
\item{markers}{A tibble of markers and allele frequencies like \code{markers40}. It
should have the columns Chrom, Locus, Pos, Allele, LocIdx, AlleIdx, and Freq,
AND it must have been run through \code{reindex_markers()}.}

\item{genome}{A tibble like that in the
\code{chromo_lengths} component of the list returned by \code{geometric_chromo_lengths()}. It
must have the columns idx, chrom, scaled_length, num_bases.}
}
\value{
This returns a tibble that is of the same format as \code{markers}, but the positions
of the markers have been simulated uniformly into the length of the genome in \code{genome}.
Note that the \code{Locus} column in \code{markers} is retained as the name of each locus.
}
\description{
If you have a bunch of markers, but you don't know where in the genome
they are, BUT you have an idea of how big the genome is and how many
chromosomes are in it, this will place your markers uniformly within
the length of the genome.  This will let you assess how much of an effect
the physical linkage might have on pairwise relationship inference.
}
\examples{
# make a fake genome
genome = geometric_chromo_lengths(n = 41, L = 7.4, sl = 0.2)$chrom_lengths
# get the markers to sprinkle into that genome
markers = markers40
}
