% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mendel-interface.R
\name{write_all_mendel_files}
\alias{write_all_mendel_files}
\title{Write all necessary files for Mendel to do a gene-dropping run}
\usage{
write_all_mendel_files(ID, Reps, Seed, Markers, Pedigree, Dir = ".")
}
\arguments{
\item{ID}{The ID or identifier of the run.  This string will be prepended to all input and output files.}

\item{Reps}{The number of replicate data sets to simulate.}

\item{Seed}{The random seed to use.  This must be a single integer between 1 and 30,000.}

\item{Markers}{A data frame in the format of \code{\link{long_markers}} that gives information
about marker positions and allele frequencies.}

\item{Pedigree}{A data frame formatted like any component of \code{\link{pedigrees}} which holds the
pedigree to drop genes upon.}

\item{Dir}{The directory in which to write the files.  It must be created already.  Defaults to the
current working directory.}
}
\description{
Writes all files in the directory Dir but does not launch mendel.
}
\examples{
# make a temp directory to place things into:
tmpDir <- tempdir()
write_all_mendel_files("mendel-example", 10, 1234, long_markers, pedigrees$HS, Dir = tmpDir)
}

