% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/break_up_LTpairs.R
\name{break_up_LTpairs}
\alias{break_up_LTpairs}
\title{break up (or augment) LTpairs into chunks of size reps}
\usage{
break_up_LTpairs(L, R)
}
\arguments{
\item{L}{a list with elements that are the LT_pairs.  Each element is a list
with elements r and c.}

\item{R}{the number of reps desired.  Should typically be about 50,000 to make
the memory footprint of each simulation manageable.}
}
\description{
This is a weird little function that lets us break up the LTpairs
that will be used for the miss_mask_mats in \code{simulate_missing_data_array()}.
The idea is that we can get extra parellelization and keep the memory overhead
for each job down low by restricting the number of reps done.  If the names
of the LTPairs list are like c("120", "121",...), the the new names will
be like c("120_1", "120_2", "120_3", 121_1", "121_2", etc...)
}
