% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{brook_trout_genos}
\alias{brook_trout_genos}
\title{Microsatellite data from brook trout}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 20262 rows and 11 columns.
}
\usage{
brook_trout_genos
}
\description{
These are genotypes of 307 brook trout from a study by Ruzzante et al.
in 2019. Each fish was typed at 33 microsatellite loci.  Missing data
are denoted by NAs.
The columns are: Population, age_class, Sample_ID, Length,
Age, Chrom, Locus, Pos, gene_copy, Allele.
}
\keyword{datasets}
