% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CKMRsim-package.R
\docType{package}
\name{CKMRsim}
\alias{CKMRsim-package}
\alias{CKMRsim}
\title{My package documentation}
\description{
bla bla, need to write more here
}
\author{
\strong{Maintainer}: Eric C. Anderson \email{eric.anderson@noaa.gov}

}
