% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_ckmr.R
\name{create_ckmr}
\alias{create_ckmr}
\title{prepare to simulate pairwise relationships from locus data}
\usage{
create_ckmr(
  D,
  kappa_matrix = kappas[c("MZ", "PO", "FS", "HS", "U"), ],
  ge_mod_assumed = ge_model_TGIE,
  ge_mod_true = ge_model_TGIE,
  ge_mod_assumed_pars_list = list(epsilon = 0.1),
  ge_mod_true_pars_list = list(epsilon = 0.1)
)
}
\arguments{
\item{D}{a data frame in the format of \code{\link{long_markers}}.  Monomorphic loci should have
been removed and the whole thing run through \code{\link{reindex_markers}} before passing it
to this function.}

\item{kappa_matrix}{A matrix like that in the supplied data \code{\link{kappas}}.  It should have
three columns, and rownames giving the abbreviated name of the pairwise relationship.  Each row
is a three-vector which are the Cotterman coefficients for the relationship. The first element is the
probability that the pair shares 0 gene copies IBD, the second is the prob that they share 1 gene
copy IBD, and the third is the prob that they share 2 gene copies IBD, all assuming no inbreeding.
By default, this just uses the "MZ", "PO", "FS", "HS", and "U" rows from \code{\link{kappas}}.}

\item{ge_mod_assumed}{The genotyping error model assumed for the analysis.}

\item{ge_mod_true}{The actual, "true" genotyping error model for the simulation.}

\item{ge_mod_assumed_pars_list}{a named list of extra arguments (besides L)
to be passed to the ge_mod_assumed function.  Set it to NULL to use the default
values for the genotyping error model.}

\item{ge_mod_true_pars_list}{a named list of extra arguments (besides L)
to be passed to the ge_mod_true function. Set to NULL to use defaults.}
}
\description{
Starting from a data frame of marker information like that
in \code{\link{microhaps}} or \code{\link{long_markers}}. This
function takes care of all the calculations necessary to simulate
log-likelihood ratios of genotypes given different pairwise relationships.
}
