% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_L_biallelic}
\alias{example_L_biallelic}
\title{An example biallelic locus}
\format{An object of class \code{list} of length 2.}
\usage{
example_L_biallelic
}
\description{
A simple example biallelic locus with made up allele frequencies.
This is a list with two components, each of them vectors:
\itemize{
\item{freqs: }{ The frequencies of the different alleles.  The names of the alleles
are just "A" and "B"}
\item{geno_freqs: }{ The expected frequencies of the possible genotypes
given the allele freqs.  The names are the names of the genotypes, which
are the allele lengths separate by " / ".}
}
}
\keyword{datasets}
