% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_L_microhap}
\alias{example_L_microhap}
\title{An example microhaplotype locus}
\format{
An object of class \code{list} of length 2.
}
\usage{
example_L_microhap
}
\description{
A simple example microhaplotype locus with three SNPs and six alleles.
This was taken from a rockfish data that we had in the lab.
This is a list with two components, each of them vectors:
\itemize{
\item{freqs: }{ The frequencies of the different alleles.  The names of
give the sequence of SNP allele at the microhaplotype.}
\item{geno_freqs: }{ The expected frequencies of the possible genotypes
given the allele freqs.  The names are the names of the genotypes, which
are the allele names separate by " / ".}
}
}
\keyword{datasets}
