% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckmr-class.R
\name{format.ckmr}
\alias{format.ckmr}
\title{format method for ckmr class (to print)}
\usage{
\method{format}{ckmr}(C)
}
\arguments{
\item{C}{an object of class \code{\link{ckmr_class}}.}
}
\description{
Just prints relevant information for a quick look.
}

