% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Qij-class.R
\name{format.Qij}
\alias{format.Qij}
\title{format method for Qij class (to print)}
\usage{
\method{format}{Qij}(Q)
}
\arguments{
\item{Q}{an object of class \code{\link{Qij_class}}.}
}
\description{
Just prints relevant information for a quick look.
}
