% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kappas}
\alias{kappas}
\title{A matrix of Cotterman coefficients for a variety of pairwise relationships}
\format{A matrix.  The rows are relationships (see the CKMRsim paper for abbreviations).
The columns are the kappas.}
\usage{
kappas
}
\description{
A matrix with named rows (relationships) and named columns (the kappas).
Eventually each row should go into a list along with the pedigree information
as well.  But for now I wanted to get it down somewhere.
}
\keyword{datasets}

