% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{markers40}
\alias{markers40}
\title{40 dinucleotide markers for a quick test data set}
\format{\preformatted{Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	80 obs. of  7 variables:
 $ Chrom  : int  4 4 4 4 4 4 4 4 4 4 ...
 $ Locus  : chr  "chr4-192" "chr4-192" "chr4-301" "chr4-301" ...
 $ Pos    : num  9967697 9967697 15854255 15854255 23048860 ...
 $ Allele : chr  "a2" "a7" "a7" "a1" ...
 $ LocIdx : int  1 1 2 2 3 3 4 4 5 5 ...
 $ AlleIdx: int  1 2 1 2 1 2 1 2 1 2 ...
 $ Freq   : num  0.846 0.154 0.75 0.25 0.64 ...
}}
\usage{
markers40
}
\description{
This was made by running the example in \code{\link{reindix_markers}}.
See \code{\link{long_markers}} for more info.
}
\keyword{datasets}

