% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{markers40}
\alias{markers40}
\title{40 dinucleotide markers for a quick test data set}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 80 rows and 7 columns.}
\usage{
markers40
}
\description{
This was made by running the example in \code{\link{reindix_markers}}.
See \code{\link{long_markers}} for more info.
}
\keyword{datasets}
