% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{microhaps}
\alias{microhaps}
\title{100 microhaplotype markers for a quick test data set}
\format{\preformatted{Classes ‘tbl_df’, ‘tbl’ and 'data.frame':	378 obs. of  7 variables:
 $ Chrom  : chr  "ddRAD" "ddRAD" "ddRAD" "ddRAD" ...
 $ Locus  : chr  "ddRAD_1" "ddRAD_1" "ddRAD_2" "ddRAD_2" ...
 $ Pos    : int  1 1 2 2 3 3 4 4 5 5 ...
 $ Allele : chr  "A" "G" "C" "T" ...
 $ LocIdx : int  1 1 2 2 3 3 4 4 5 5 ...
 $ AlleIdx: int  1 2 1 2 1 2 1 2 1 2 ...
 $ Freq   : num  0.99 0.01 0.995 0.005 0.705 0.295 0.705 0.295 0.995 0.005 ...
}}
\usage{
microhaps
}
\description{
This was made by running some code in scratch/create-data.R.
See \code{\link{long_markers}} for more info about the columns.
Note that Chrom is a character here (since they are unmapped) and
are all the same.  Pos is actually ordered the way I want them to be.
That turns out to be necessary for proper sorting with \code{\link{reindex_markers}}.
}
\keyword{datasets}

