% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_Qij_with_missing_data.R
\name{model_Qij_with_missing_data}
\alias{model_Qij_with_missing_data}
\title{Model missing-data simulation results to interpolate FNR and FPR values}
\usage{
model_Qij_with_missing_data(
  Qtib,
  nu,
  de = "U",
  tr = "U",
  FNRs = c(0.3, 0.2, 0.1, 0.05, 0.01)
)
}
\arguments{
\item{Qtib}{A tibble of missing-data simulation results, typically assembled
from the Snakemake outputs created by \code{summarize_missing_data()}. It must
include \code{num_non_missing_loci}, \code{num_missing_loci}, \code{Qijs_unlinked}, and
\code{Qijs_linked}.}

\item{nu}{the name of the relationship that is in the \strong{nu}merator
of the likelihood ratio (Lambda) whose distribution you wish to learn about.
It is a string, for example "FS", or "PO", or "U".  The Q values for that
relationship must be included in parameter Q.  If this is a vector, then
all different values are used in combination with all the values of
\code{de}, \code{tr}, and, possibly, \code{pstar}. Corresponds to column "numerator" in the output}

\item{de}{the relationship that appears in the \strong{de}nominator of Lambda.
By default it is "U".  Corresponds to column "denominator" in the output. If it
is a vector, then all values are done iteratively in combination with other values as
described for \code{nu}.}

\item{tr}{the true relationship of the pairs. Default is "U". (i.e. you are going to
get samples of Lambda under their distribution when the true relationship is tr).
Operates over all values if a vector. Corresponds to column "true_relat" in the
output.}

\item{FNRs}{the false negative rates at which to evaluate the false positive rates.
These are reported in column "fnr" in the output. These should all be between
0 and 1.  By default fnr is c(0.3, 0.2, 0.1, 0.05, 0.01, 0.001).}
}
\description{
More later.
}
\examples{
# here for testing at the moment
#Qtib <- readr::read_rds("/tmp/Qtib.rds")

}
