% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ckmr-class.R
\name{print.ckmr}
\alias{print.ckmr}
\title{print method for ckmr class}
\usage{
\method{print}{ckmr}(x, ...)
}
\arguments{
\item{x}{an object of class \code{\link{ckmr_class}}.}

\item{...}{additional arguments to print. (But nothing implemented).}
}
\description{
Just wraps a call to the format.ckmr function
}
