% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mendel-interface.R
\name{run_mendel}
\alias{run_mendel}
\title{Run the mendel binary in directory Dir using control file Control}
\usage{
run_mendel(Dir, Control)
}
\arguments{
\item{Dir}{the directory in which to run mendel}

\item{Control}{the "control file" with which to run mendel.  If not an absolute path
it should be given relative to the Dir directory.}
}
\description{
Run the mendel binary in directory Dir using control file Control
}
\examples{
# first prepare the input files and define tmpDir
example(write_all_mendel_files)

# then run mendel
run_mendel(tmpDir, "mendel-example-Control.in")
}
