% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate_and_calc_Q}
\alias{simulate_and_calc_Q}
\title{compute all-locus genotype pair probs for pairs simulated given specified relationships}
\usage{
simulate_and_calc_Q(YL, reps = 10^4, froms = NULL, tos = NULL)
}
\arguments{
\item{YL}{the list that holds the Y_l and Y_l_true matrices}

\item{reps}{number of pairs to simulate for each relationship in froms.  This recycles
to be the same length as froms, but you can specify different numbers of reps for each
of the relationships.}

\item{froms}{a vector of names of the relationship IDs to simulate from. If this is NULL then
the function will just simulate from all the values that have had Y_l_true matrices computed for them
in the first component of YL.  Genotype values get simulated from the Y_l_true values}

\item{tos}{a vector of names of the relationship IDs to calculate the genotype log probs of the simulated
genotypes from.  Genotype log probs are calculated using the Y_l matrices. If this is NULL then
the function will just compute probs for all the relationships that have had Y_l matrices computed for them
in YL.}
}
\value{
This returns a list with components that are the relationships that were simulated
from.  Inside each of those components is a list with components referring to the relationships
that had genotype log probabilities calculated (the "tos"). The contents of each is a vector
of length reps which are the log probability of the multilocus genotypes of that each of reps
simulated pairs.
}
\description{
Once the Y_l and Y_l_true matrices have been inserted into the list of locus computations, then
it is time to simulate multilocus genotype pairs under a variety of relationships,
and then compute the log probability of them assuming various relationships.
}
\examples{
example(insert_Y_l_matrices)
mh_sacQ_example <- simulate_and_calc_Q(mh_yl_example, 10^4, froms = c("U", "PO"), tos = c("U", "PO"))

\dontrun{
# and we could plot the distribution of those if we wanted to
m <- mh_sacQ_example
D <- data.frame(U = m$U$PO - m$U$U, PO = m$PO$PO - m$PO$U)
DD <- tidyr::gather(D, Relat, logL)
library(ggplot2)
ggplot(DD, aes(x = logL, fill = Relat)) + geom_density(alpha = 0.6)
}
}

