% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_missing_data_array.R
\name{simulate_missing_data_array}
\alias{simulate_missing_data_array}
\title{Simulate Qijs across multiple relevant levels of missing data}
\usage{
simulate_missing_data_array(
  input_path,
  num_points = 11,
  tabulate_and_exit = FALSE,
  simulation_approach = "simple_miss_freq",
  ...,
  .parallel_plan = "sequential",
  .workers = NULL,
  CKMRsim.discard_stderr = TRUE,
  CKMRsim.discard_stdout = TRUE
)
}
\arguments{
\item{input_path}{the absolute path to the file of inputs written out by
\code{summarize_missing_data()}.}

\item{num_points}{Number of points between the observed min and max non-missing loci to simulate.}

\item{simulation_approach}{this determines how simulations should be done. The
options are:
\itemize{
\item \code{simple_miss_freq}: do simulations at \code{num_points} different values of the number
of pairwise shared genotypes, drawing the pattern of missingness from a very simple
independent random sampling model according to the frequency of missing data in all
loci.
\item \code{pairwise_patterns}: do a separate simulation for each number of shared non-missing
loci observed amongst all pairs.  The simulation will do at least one rep for each
pair within a category of the number of shared non-missing loci, and it will use
the observed patterns of missing loci for all pairs. If the number Np of pairs in a category
is less than \code{reps}, then the actual number of reps done will be the smallest integer
multiple of Np >= \code{reps}, and the patterns of missing data will be recycled.
}}

\item{...}{
  Arguments passed on to \code{\link[=simulate_Qij]{simulate_Qij}}
  \describe{
    \item{\code{sim_relats}}{a vector of names of the relationship IDs (these were the rownames
in the \code{kappa_matrix} argument to \code{\link{create_ckmr}} to simulate from.
For each relationship ID in sim_relats, genotype values will get simulated
from the \code{Y_l_true} values in \code{C}.}
    \item{\code{calc_relats}}{Either a vector of names of the relationship IDs to calculate the
genotype log probabilities of the simulated
genotypes for,  or a (named) list with one component for each of the names in sim_relats
that is, itself, a vector of relationships that you want the probability calculated for.
Genotype log probs are calculated using the Y_l matrices.}
    \item{\code{reps}}{a synonym for \code{calc_relats} for compatibility to an earlier version
of CKMRsim.}
  }}

\item{.parallel_plan}{Future strategy (e.g., "sequential", "multisession").}

\item{.workers}{Number of workers for parallel backend.}

\item{CKMRsim.discard_stderr}{logical.  Set to true by default to discard stderr,}

\item{CKMRsim.discard_stdout}{logical. Set to true by default to discard stdout}
}
\description{
This function helps estimate error rates (FPRs/FNRs) under realistic missing data.
Some data sets have a lot of missing markers.  If this is the case, it is not
OK to just do the simulations as if there is no missing data.  This function
wraps up a lot of different steps that can be taken to try to get more accurate
"first-pass" FPRs and FNRs for situations with a lot of missing data.
}
\details{
Note that this function requires a path to the output of \code{summarize_missing_data()}.
Rather than passing things in directly, we write an intermediate, uncompressed
file to a temp directory of the necessary inputs so that we can easily parallelize
this on a cluster if needed.

Steps:
\enumerate{
\item Summarize missingness across individuals and pairs
\item Estimate per-locus missing rates, then derive pairwise missing probabilities.
\item Record background stats and plots
\item Clean and prepare dot args
\item Simulate Qij values using either random missingness (when \code{simulation_approach == "simple_miss_freq"})
or observed pairwise patterns of shared loci (when \code{simulation_approach == "pairwise_patterns"}).
}
}
