% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_missing_data_array.R
\name{simulate_missing_data_array}
\alias{simulate_missing_data_array}
\title{Simulate Qijs across multiple relevant levels of missing data}
\usage{
simulate_missing_data_array(LG, C, num_points = 11, num_cores = 1, ...)
}
\arguments{
\item{LG}{the genotypes in long format.  It must have the columns
Indiv (unique IDs of the individuals), Locus, gene_copy (must be 1 or 2
denoting which of the two gene copies in a diploid each allele is), and
Allele, which must be a character. If there are any missing genotypes
in the data frame, they must appear as NAs in the Allele column.}

\item{C}{the ckmr object upon which to base the simulations.}

\item{num_points}{the number of different values between the lowest observed
number of pairwise non-missing genotypes and the highest, inclusive, that
simulations will be performed for.}

\item{num_cores}{Number of cores to parallelize the simulations over (using mclapply)
from the parallel package.  On Windows, parallelization is not available from forking
so this must remain equal to 1 on Windows.}

\item{...}{
  Arguments passed on to \code{\link[=simulate_Qij]{simulate_Qij}}
  \describe{
    \item{\code{sim_relats}}{a vector of names of the relationship IDs (these were the rownames
in the \code{kappa_matrix} argument to \code{\link{create_ckmr}} to simulate from.
For each relationship ID in sim_relats, genotype values will get simulated
from the \code{Y_l_true} values in \code{C}.}
    \item{\code{calc_relats}}{a vector of names of the relationship IDs to calculate the
genotype log probabilities of the simulated
genotypes from.  Genotype log probs are calculated using the Y_l matrices.}
    \item{\code{reps}}{a synonym for \code{calc_relats} for compatibility to an earlier version
of CKMRsim.}
    \item{\code{unlinked}}{A logical indicating whether to simulate the markers as unlinked.  By default this is
TRUE.  If FALSE, then genotypes at linked markers will be simulated using the program MENDEL, genotyping
errors will be applied to them, and the Q_ij values themselves will still be computed under the assumption
of no linkage. However, they will be simulated under the no-linkage model for relationships "U", "PO", and
"MZ", because, in the absence of LD, related pairs under those relationships are not affected by
physical linkage.}
    \item{\code{pedigree_list}}{If you specify unlinked == FALSE, then you have to supply a pedigree_list.}
  }}
}
\value{
This function returns a list.  More on that later.
}
\description{
Some data sets have a lot of missing markers.  If this is the case, it is not
OK to just do the simulations as if there is no missing data.  This function
wraps up a lot of different steps that can be taken to try to get more accurate
"first-pass" FPRs and FNRs for situations with a lot of missing data.
The steps are:
\itemize{
\item Tabulate the distribution of the number of informative (i.e., not missing in
either member of the pair) markers, across all pairs. (Note, this requires
that you have an actual data set that you are trying to do relationship
inference in.)
\item Estimate missingness rates per locus, and from that calculate the rate of
missingness in pairs, under a simple independence assumption.
\item Simulate Q_ij values at a series of different numbers of non-missing loci
to calculate FPRs and FNRs for those.
}
}
\examples{
# this is just here for testing at the moment
LG <- read_rds("/tmp/LG.rds")
C <- read_rds("/tmp/C.rds")
}
