% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_Qij.R
\name{simulate_Qij}
\alias{simulate_Qij}
\title{compute the all-locus genotype pair probs for pairs simulated given specified relationships}
\usage{
simulate_Qij(
  C,
  sim_relats,
  calc_relats,
  reps = 10^4,
  unlinked = TRUE,
  forceLinkagePO = FALSE,
  pedigree_list = NULL,
  useMendel = FALSE,
  cM_per_Mb = 1,
  min_crossovers = 0,
  miss_mask_mat = NULL,
  rando_miss_wts = NULL,
  rando_miss_n = 0,
  froms,
  tos
)
}
\arguments{
\item{C}{the ckmr object upon which to base the simulations.}

\item{sim_relats}{a vector of names of the relationship IDs (these were the rownames
in the \code{kappa_matrix} argument to \code{\link{create_ckmr}} to simulate from.
For each relationship ID in sim_relats, genotype values will get simulated
from the \code{Y_l_true} values in \code{C}.}

\item{calc_relats}{Either a vector of names of the relationship IDs to calculate the
genotype log probabilities of the simulated
genotypes for,  or a (named) list with one component for each of the names in sim_relats
that is, itself, a vector of relationships that you want the probability calculated for.
Genotype log probs are calculated using the Y_l matrices.}

\item{reps}{a synonym for \code{calc_relats} for compatibility to an earlier version
of CKMRsim.}

\item{unlinked}{A logical indicating whether to simulate the markers as unlinked.  By default this is
TRUE.  If FALSE, then genotypes at linked markers will be simulated using CKMRsim's
internal Rcpp simulator, or using the program MENDEL if \code{useMendel = TRUE}; genotyping
errors will be applied to them, and the Q_ij values themselves will still be computed under the assumption
of no linkage. However, they will be simulated under the no-linkage model for relationships "U", "PO", and
"MZ", because, in the absence of LD, related pairs under those relationships are not affected by
physical linkage.}

\item{forceLinkagePO}{If you really want to force simulation to be done under physical linkage for the PO case
(perhaps to verify that you get the same result as with unlinked.  Pass TRUE to this while \code{unlinked} is
FALSE.)}

\item{pedigree_list}{If you specify unlinked == FALSE, then you have to supply a pedigree_list.}

\item{useMendel}{If TRUE, use the external Mendel program for linked gene
dropping. If FALSE, use CKMRsim's internal Rcpp simulator.}

\item{cM_per_Mb}{Recombination-rate conversion used to create a Map column
from Pos when Map is not present. The value is centiMorgans per megabase.}

\item{min_crossovers}{Minimum number of crossovers in the typed marker span
of each chromosome for each meiosis.}

\item{miss_mask_mat}{A logical matrix with length(YL) columns and reps rows.  The (r,c)-th is TRUE if
the c-th locus should be considered missing in the r-th simulated sample.  This type of specification
lets the user simulate either a specific pattern of missingness, if desired, or to simulate patterns of
missing data given missing data rates, etc.}

\item{rando_miss_wts}{weights to be given to different loci that influence whether they
will be one of the rando_miss_n missing loci in any iteration.  These will be recycled
(or truncated) to have length equal to the number of loci, and they will be normalized
to sum to one as appropriate (so you can provide
them in unnormalized form.) The idea of this is to be able to use observed rates of
missingness amongst loci to mask some loci as missing.  Given as a comma-delimited
string in column "rando_miss_wts" in the output.}

\item{rando_miss_n}{a single number less than the number of loci.  Each iteration,
rando_miss_n loci will be considered missing, according to the rando_miss_wts.  This
let's you get a sense for how well you will do, on average, with a certain number of
missing loci.}

\item{froms}{a synonym for \code{sim_relats} for compatibility to an earlier version
of CKMRsim.}

\item{tos}{a synonym for \code{calc_relats} for compatibility to an earlier version
of CKMRsim.}
}
\description{
After you have created a ckmr object with \code{\link{create_ckmr}}, then
it is time to simulate multilocus genotype pairs under all the
relationships that you want to simulate from, and compute the likelihood of those
relationships under different relationship hypotheses.
This function does that.
}
