% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_Qij.R
\name{simulate_Qij}
\alias{simulate_Qij}
\title{compute the all-locus genotype pair probs for pairs simulated given specified relationships}
\usage{
simulate_Qij(C, froms, tos, reps = 10^4, unlinked = TRUE,
  forceLinkagePO = FALSE)
}
\arguments{
\item{C}{the ckmr object upon which to base the simulations.}

\item{froms}{a vector of names of the relationship IDs to simulate from. For each relationship ID in
froms, genotype values will get simulated from the Y_l_true values in C.}

\item{tos}{a vector of names of the relationship IDs to calculate the genotype log probs of the simulated
genotypes from.  Genotype log probs are calculated using the Y_l matrices.
#' @param reps An integer. The number of pairs to simulate for each relationship in \code{froms}.
Default is 10,000.}

\item{unlinked}{A logical indicating whether to simulate the markers as unlinked.  By default this is
TRUE.  If FALSE, then genotypes at linked markers will be simulated using the program MENDEL, genotyping
errors will be applied to them, and the Q_ij values themselves will still be computed under the assumption
of no linkage. However, they will be simulated under the no-linkage model for relationships "U", "PO", and
"MZ", because, in the absence of LD, related pairs under those relationships are not affected by
physical linkage.}

\item{forceLinkagePO}{If you really want to force simulation to be done under physical linkage for the PO case
(perhaps to verify that you get the same result as with unlinked.  Pass TRUE to this while \code{unlinked} is
FALSE.)}
}
\description{
After you have created a ckmr object with \code{\link{create_ckmr}}, then
it is time to simulate multilocus genotype pairs under all the
relationships that you want to simulate from, and compute the likelihood of those
relationships under different relationship hypotheses.
This function does that.
}

