% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_missing_data.R
\name{summarize_missing_data}
\alias{summarize_missing_data}
\title{Summarize the missing data and make some plots}
\usage{
summarize_missing_data(LG, C, snakemake_dir = NA, snake_rep_split = 50000)
}
\arguments{
\item{LG}{the genotypes in long format.  It must have the columns
Indiv (unique IDs of the individuals), Locus, gene_copy (must be 1 or 2
denoting which of the two gene copies in a diploid each allele is), and
Allele, which must be a character. If there are any missing genotypes
in the data frame, they must appear as NAs in the Allele column.}

\item{C}{A CKMR object created by \code{create_ckmr()}.}

\item{snakemake_dir}{Name of directory to create in order to write out the
materials for running the simulations via a Snakefile.}

\item{snake_rep_split}{If snakemake_dir is non-NA, this is the number of simulation
reps to be done for each snakemake job.  If a partition has fewer pairs that snake_rep_split
of if snake_rep_split is not perfectly divisible by the number of pairs
then the pairs are recycled so that every job has exactly snake_rep_split reps.}
}
\value{
A list containing summary plots and statistics, or the absolute path
to the Snakemake directory when \code{snakemake_dir} is provided.
}
\description{
In default mode, this will just summarize the missing data and make and
store some plots in the list output. If you provide \code{snakemake_dir}, this
function writes a self-contained Snakemake arena for running the missing-data
simulations in separate R processes.
}
\details{
Note that C must be a CKMR object that is suitable for \emph{both} linked and unlinked
simulation.  So, if you don't have that, then you best
prepare it, even if it means sprinkling your markers into a pseudogenome
as described \href{https://eriqande.github.io/tws-ckmr-2022/kin-finding-lab.html#power-for-kin-finding-while-accounting-for-physical-linkage}{here}.
}
