% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_missing_data.R
\name{summarize_missing_data}
\alias{summarize_missing_data}
\title{Summarize the missing data and make some plots.}
\usage{
summarize_missing_data(
  LG,
  C,
  downstream_tempdir = NA,
  snakemake_dir = NA,
  snake_rep_split = 50000
)
}
\arguments{
\item{LG}{the genotypes in long format.  It must have the columns
Indiv (unique IDs of the individuals), Locus, gene_copy (must be 1 or 2
denoting which of the two gene copies in a diploid each allele is), and
Allele, which must be a character. If there are any missing genotypes
in the data frame, they must appear as NAs in the Allele column.}

\item{downstream_tempdir}{By default, this is NA.  If not NA, then the function will
write the necessary objects for \code{simulate_missing_data_array()} into a temp file
within this temp dir, and it will return the path to that file (to be passed into
\code{simulate_missing_data_array()}).  If you are working in a cluster environment on
different nodes, make sure that \code{downstream_tempdir} points to storage that is
accessible from all nodes (for example, a directory in scratch, etc.). If you set
downstream_tempdir to NULL, then the function will try to use the environment variables
CKMRSIM_SHARED_TMPDIR and TMPDIR, in that order, before just using a standard
R temp directory}

\item{snakemake_dir}{Name of directory to create in order to write out the
materials for running the simulations via a Snakefile.}

\item{snake_rep_split}{If snakemake_dir is non-NA, this is the number of simulation
reps to be done for each snakemake job.  If a partition has fewer pairs that snake_rep_split
of if snake_rep_split is not perfectly divisible by the number of pairs
then the pairs are recycled so that every job has exactly snake_rep_split reps.}
}
\value{
A list containing summary plots, statistics, and simulation results, or
the absolute path to a tempfile with inputs needed for \code{simulate_missing_data_array()}.
}
\description{
In default mode, this will just summarize the missing data and make and
store some plots in the list output.  If you provide a temp directory
path for \code{downstream_tempdir}, then this function will assemble all that
is needed to run \code{simulate_missing_data_array()} and will write that out
to a tempfile in that temp directory so that it can be accessed by different
workers from within a \code{future_lapply()} call so that this can be parallelized
across a cluster.
}
\details{
Note that C must be a CKMR object that is suitable for \emph{both} linked and unlinked
simulation.  So, if you don't have that, then you best
prepare it, even if it means sprinkling your markers into a pseudogenome
as described \href{https://eriqande.github.io/tws-ckmr-2022/kin-finding-lab.html#power-for-kin-finding-while-accounting-for-physical-linkage}{here}.
}
